+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Added
+- S3 bucket cleaner.
+
 ## [1.0.0] - 2022-12-09
  
 ### Added